+++ Rakefile
@@ -6,19 +6,19 @@ end
 
 namespace :fetch do
   task :timpul do
-    TimpulFetcher.new.run
+    Fetchers::Timpul.new.run
   end
 
   task :publika do
-    PublikaFetcher.new.run
+    Fetchers::Publika.new.run
   end
 
   task :unimedia do
-    UnimediaFetcher.new.run
+    Fetchers::Unimedia.new.run
   end
 
   task :protv do
-    ProTvFetcher.new.run
+    Fetchers::ProTv.new.run
   end
 
   task :agora do
@@ -45,7 +45,7 @@ namespace :watch do
   task :timpul do
     while true do
       puts "Restarting Fetcher"
-      TimpulFetcher.new.run
+      Fetchers::Timpul.new.run
       puts "Restarting parser"
       TimpulParser.new.run
       sleep 10
@@ -55,7 +55,7 @@ namespace :watch do
   task :publika do
     while true do
       puts "Restarting Fetcher"
-      PublikaFetcher.new.run
+      Fetchers::Publika.new.run
       puts "Restarting parser"
       PublikaParser.new.run
       sleep 10
@@ -65,7 +65,7 @@ namespace :watch do
   task :unimedia do
     while true do
       puts "Restarting Fetcher"
-      UnimediaFetcher.new.run
+      Fetchers::Unimedia.new.run
       puts "Restarting parser"
       UnimediaParser.new.run
       sleep 10
