+++ lib/local_storage.rb
