+++ main.rb
@@ -15,6 +15,8 @@ class Settings < Settingslogic
   source "./config.yml"
 end
 
+require_relative "lib/local_storage"
+
 require_relative "lib/person"
 require_relative "lib/smart_fetcher"
 require_relative "lib/parsed_page"
@@ -22,9 +24,6 @@ require_relative "lib/word"
 require_relative "lib/batch_racai_fetcher"
 require_relative "lib/racai_builder"
 require_relative "analyzer"
-
-# fetcher helper classes
-require_relative "fetchers/local_storage"
 
 # fetcher implementations
 require_relative "fetchers/timpul"
