--- scripts/fetch.rb
@@ -1,5 +1,0 @@
-require_relative "../main"
-
-PublikaFetcher.new.run
-UnimediaFetcher.new.run
-TimpulFetcher.new.run
